% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{big.a}
\alias{big.a}
\title{Helper Function computes truncated sum of inverse squares}
\usage{
big.a(n)
}
\arguments{
\item{n}{where to truncate the sum}
}
\value{
the truncated sum
}
\description{
Comment
}
