% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{f.FuLiF.max}
\alias{f.FuLiF.max}
\title{Helper Function}
\usage{
f.FuLiF.max(n, S, j, k)
}
\arguments{
\item{n}{comment}

\item{S}{comment}

\item{j}{comment}

\item{k}{comment}
}
\value{
the function
}
\description{
Comment
}
