% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{getCommand}
\alias{getCommand}
\title{Helper Function}
\usage{
getCommand(n, S, theta, scenario = "neutral", ...)
}
\arguments{
\item{n}{comment}

\item{S}{comment}

\item{theta}{comment}

\item{scenario}{comment}

\item{...}{comment}
}
\value{
the function
}
\description{
Comment
}
