% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{little.a}
\alias{little.a}
\title{Helper Function computes truncated harmonic sum}
\usage{
little.a(n)
}
\arguments{
\item{n}{where to truncate the sum}
}
\value{
the truncated harmonic sum
}
\description{
Comment
}
