% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{plot_Fu_li_n}
\alias{plot_Fu_li_n}
\title{Helper Function}
\usage{
plot_Fu_li_n(n.lower, n.upper, S, j, k, ...)
}
\arguments{
\item{n.lower}{comment}

\item{n.upper}{comment}

\item{S}{comment}

\item{j}{comment}

\item{k}{comment}

\item{...}{comment}
}
\value{
the function
}
\description{
Comment
}
