% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{plot_Fu_li_S}
\alias{plot_Fu_li_S}
\title{Helper Function}
\usage{
plot_Fu_li_S(n, j, k, ...)
}
\arguments{
\item{n}{comment}

\item{j}{comment}

\item{k}{comment}

\item{...}{comment}
}
\value{
the function
}
\description{
Comment
}
