% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{plot_n}
\alias{plot_n}
\title{A function to plot model-independent bounds on summary statistics
as they depend on the number of individuals sampled.}
\usage{
plot_n(f = "FuLi", n.lower, n.upper, S, j, k, ...)
}
\arguments{
\item{f}{a string denoting statistic {FuLi, TajPi, TajD}}

\item{n.lower}{the minimum number of sampled individuals, always set >=3}

\item{n.upper}{the maximum number of sampled individuals}

\item{S}{the number of segregating sites or polymorphic sites}

\item{j}{the argument of the component of the site frequency spectrum}

\item{k}{the component of the site frequency spectrum}

\item{...}{additional arguments to be passed (none in actuality)}
}
\value{
plot of bounds according to parameter inputs
}
\description{
Return a plot specified by the input parameters.
}
\examples{
plot_n(f="FuLi",n.lower=3,n.upper=50,S=100,j=1,k=5)

}
\keyword{individuals}
\keyword{plot,}
\keyword{sampled}
