% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{plot_S}
\alias{plot_S}
\title{A function to plot model-independent bounds on summary statistics
as they depend on the number of segregating or polymorphic sites.}
\usage{
plot_S(f = "FuLi", n, j, k, ...)
}
\arguments{
\item{f}{a string denoting statistic {FuLi, TajPi, TajD}}

\item{n}{the number of sampled individuals}

\item{j}{the argument of the component of the site frequency spectrum}

\item{k}{the component of the site frequency spectrum}

\item{...}{additional arguments to be passed (none in actuality)}
}
\value{
plot of bounds according to parameter inputs
}
\description{
Return a plot specified by the input parameters.
}
\examples{
plot_S(f="FuLi",n=50,j=1,k=5)

}
\keyword{individuals}
\keyword{plot,}
\keyword{sampled}
