% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{plot_statistic}
\alias{plot_statistic}
\title{A function to generate and plot summary statistics of parametric molecular genetic data,
together with their model-independent bounds.}
\usage{
plot_statistic(f = "FuLi", R, n, S, j, theta, scenario = "neutral",
  plot_points = FALSE, ...)
}
\arguments{
\item{f}{a string denoting statistic {FuLi, TajPi, TajD}}

\item{R}{the number of data points to generate from simulation runs}

\item{n}{the number of sampled individuals}

\item{S}{the number of segregating sites or polymorphic sites}

\item{j}{the argument of the component of the site frequency spectrum}

\item{theta}{the mutation rate, need to be specified if plot_points=TRUE}

\item{scenario}{the parametric model {neutral, bottleneck, expansion}, if plot_points=TRUE}

\item{plot_points}{option of generating data from parametric model {TRUE, FALSE}}

\item{...}{additional arguments to be passed (none in actuality)}
}
\value{
plot of bounds according to parameter inputs, and optionally data points
}
\description{
Return a plot specified by the input parameters.
}
\examples{
plot_statistic(f="FuLi",R=100,n=500,S=100,j=1,theta=5.0,scenario="neutral",plot_points=FALSE)

}
\keyword{plot,}
\keyword{simulation}
