% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{plot_tajima_D_k}
\alias{plot_tajima_D_k}
\title{Helper Function}
\usage{
plot_tajima_D_k(R, n, S, j, theta, scenario, plot_points = FALSE, ...)
}
\arguments{
\item{R}{comment}

\item{n}{comment}

\item{S}{comment}

\item{j}{comment}

\item{theta}{comment}

\item{scenario}{comment}

\item{plot_points}{comment}

\item{...}{comment}
}
\value{
the function
}
\description{
Comment
}
