% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{plot_tajima_D_S}
\alias{plot_tajima_D_S}
\title{Helper Function}
\usage{
plot_tajima_D_S(n, j, k)
}
\arguments{
\item{n}{comment}

\item{j}{comment}

\item{k}{comment}
}
\value{
the function
}
\description{
Comment
}
