% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{plot_tajima_Pi_n}
\alias{plot_tajima_Pi_n}
\title{Helper Function}
\usage{
plot_tajima_Pi_n(n.lower, n.upper, S, j, k)
}
\arguments{
\item{n.lower}{comment}

\item{n.upper}{comment}

\item{S}{comment}

\item{j}{comment}

\item{k}{comment}
}
\value{
the function
}
\description{
Comment
}
