% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundsForResults.R
\name{runCommand}
\alias{runCommand}
\title{Helper Function}
\usage{
runCommand(R, n, S, theta, f = "FuLi", scenario = "neutral", ...)
}
\arguments{
\item{R}{comment}

\item{n}{comment}

\item{S}{comment}

\item{theta}{comment}

\item{f}{comment}

\item{scenario}{comment}

\item{...}{comment}
}
\value{
the function
}
\description{
Comment
}
