% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteFreqBounds.R
\docType{package}
\name{siteFreqBounds}
\alias{siteFreqBounds}
\alias{siteFreqBounds-package}
\title{Package: siteFreqBounds}
\description{
A collection of functions plotting model-independent bounds on summary statistics of the site frequency
spectrum. See examples
for functions \code{\link{f.min}} and  \code{\link{f.max}}.
}
\seealso{
\code{\link{f.min}}

\code{\link{f.max}}

\code{\link{f.thetaPi.max}}

\code{\link{f.thetaPi.min}}

\code{\link{plot_tajima_Pi_k}}

\code{\link{plot_tajima_Pi_n}}

\code{\link{plot_tajima_Pi_S}}

\code{\link{little.a}}

\code{\link{big.a}}

\code{\link{v.T}}

\code{\link{u.T}}

\code{\link{v.F}}

\code{\link{u.F}}

\code{\link{f.TajimaD.max}}

\code{\link{f.TajimaD.min}}

\code{\link{plot_tajima_D_k}}

\code{\link{plot_tajima_D_S}}

\code{\link{plot_tajima_D_n}}

\code{\link{f.FuLiF.max}}

\code{\link{f.FuLiF.min}}

\code{\link{plot_Fu_li_k}}

\code{\link{plot_statistic}}

\code{\link{plot_n}}

\code{\link{plot_S}}

\code{\link{plot_Fu_li_S}}

\code{\link{plot_Fu_li_n}}

\code{\link{getCommand}}

\code{\link{runCommand}}
}
